Moved Config, Updated Deps, Updated gitignore
