+++ package.json
@@ -22,6 +22,6 @@
 		"@types/node": "^12.12.14",
 		"@types/yargs": "^13.0.3",
 		"eslint": "^6.7.2",
-		"typescript": "^3.7.2"
+		"typescript": "^3.7.3"
 	}
 }
